// models/social.ts
export interface SocialModel {
  id?: string;
  eventDate: number;
  eventDescription: string;
  eventImage: string;
  eventLocation: string;
  eventName: string;
  currentUser: string;
  interested: boolean;

  // TODO: You may need to add attributes here
  // to implement your desired functionality.
  // The staff solutions add two attributes.
}